Agentic AI Projects/Report Generation Agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
python-dotenv==1.0.0
reportlab==4.0.7
pydantic==2.5.0
aiofiles==23.2.1
jinja2==3.1.2
pandas==2.1.3
numpy==1.26.2

